JVM: Unbox inline class suspend lambda parameters before invokeSuspend
Otherwise, their types are different, whether they are unspilled
at the beginning of invokeSuspend or after suspend call.

Before the change, parameter of suspend lambda was assumed to be boxed,
while spilled one - unboxed, which lead to CCE.

This change fixes the disparity - both the parameter and spilled
inline class are unboxed.

Because we use unboxed version of inline class in invokeSuspend,
there is no longer a need to check, whether we spilled null
at the beginning of the function before unboxing it,
since we do not unbox the parameter.

 #KT-76583 Fixed