Check if progress indicator has been canceled in the middle of CodeGen
On Android Studio, the Compose preview using KaFirCompilerFacility API
holds the read lock long time (like 700ms). The missing progress
indicator cancellation check of the CodeGen results in a freeze in the
view of user, and we got a crash report. This commit adds
`ProgressManager.checkCanceled()` to the interation of lowering each
file in the backend CodeGen.

^KT-76698