Add ProgressManager.checkCanceled() to compose lowering
IDE features like debugger expression evaluator, LiveEdit, and compose
preview rely on KaCompilerFacility API. For Jetpack Compose apps, it
runs the Compose compiler plugin lowering passes. We got a crash report
that the IDE had a UI freeze for 65 seconds. This commit adds
`ProgressManager.checkCanceled()` between the compose lowering passes to
allow the user to cancel the progress when it has the UI freezes.

^KT-76802