fixup! [FIR] Support negative information for not very trustworthy classes
Previously, the logic was to not report the warning if the
subject type comes from a library, but now we suppress the diagnostic
only if the `equals()` override itself both comes from a library
and is an override declared within a `data class`. The old algorithm
was unnecessarily permissive, whereas the new one is more accurate and
is constrained by the technical detail, not hypothetical assumptions
that some library authors don't know about `data` classes/objects.
The technical detail is that the same `FirDeclarationOrigin` can't
be both `Library` and `Synthetic`, so we can't tell generated `equals()`
from libraries apart from user-defined. Also, now we only report unsafe
exhaustiveness on `== MyObject` comparisons and check that the
problematic equals comes from the `object` variant specifically:
a `when` is unsafe if it can lead to a runtime error, which can only
happen if all the branches are `false`, and strange `(== MyObject) False`
may only come from situations where the subject is, actually, of the
type `MyObject`, but the `equals()` implementation returns `false`.

See the tests in the next commit for some examples.

And, on top of that, the old algorithm contained a mistake in how
the contract of `sealed` class's inheritors was calculated :)
Instead of taking `minOf` by them, it was essentially taking `maxOf`.

And also, naming the function `hasEqualsOverrideContract` was, in fact,
not correct as it calculates the opposite thing: if a class has an
`equals()` that doesn't follow any contract.

Additionally, notice that, unfortunately, scenarios where
`SAFE_FOR_SMART_CAST` and `TRUSTED_FOR_EXHAUSTIVENESS` act as unrelated
values are closer than we thought :( This is probably a sign that we
should to replace the enum with a pair of two properties, but maybe we
don't want to do it just yet...