[LL] AllCandidatesResolver: guard original fir from modifications
`FirCallCompleter` and `FirOverloadByLambdaReturnTypeResolver` in
particular were designed to be called only once on unresolved
code during the resolution, so they may modify the passed FIR
which is legal.
Hence, `AllCandidatesResolver` has to be very careful to not
trigger such modification on already resolved code.
During the fixing of KT-70199 there was an assumption that
the tree might be modified only during `runCompletionForCall`
with modes that analyze lambdas, but in reality it can
happen even during the `collectCandidates` execution via
`FirOverloadByLambdaReturnTypeResolver`.
The solution is to copy only potentially changeable part of the tree
to avoid modification of the original tree

^KT-76750 Fixed